# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backcasting Planner API",
    "description": "Adaptive planning system: backcast from a desired future state, plan, schedule against real calendar and capacity constraints, measure execution, and replan when reality diverges.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/goals/{goal_id}/milestones": {
      "post": {
        "tags": [
          "milestones"
        ],
        "summary": "Define Milestone",
        "description": "Pin a checkpoint on the goal's current backcasting run\n(docs/04 step 10).",
        "operationId": "defineMilestone",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "milestones"
        ],
        "summary": "List Milestones",
        "description": "The goal's checkpoints on its current run, earliest target\nfirst. Empty while no run exists.",
        "operationId": "listMilestones",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MilestoneResponse"
                  },
                  "title": "Response Listmilestones"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MilestoneCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "target_date": {
            "type": "string",
            "format": "date-time",
            "title": "Target Date"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "target_date"
        ],
        "title": "MilestoneCreate"
      },
      "MilestoneResponse": {
        "properties": {
          "milestone_id": {
            "type": "string",
            "format": "uuid",
            "title": "Milestone Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "goal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Goal Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "target_date": {
            "type": "string",
            "title": "Target Date"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "milestone_id",
          "run_id",
          "goal_id",
          "title",
          "description",
          "target_date",
          "created_at",
          "updated_at"
        ],
        "title": "MilestoneResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}